Add files via upload
The code allows users to input the name of an animal, fetches data about that animal from an API, serializes the data into HTML format, and generates an HTML webpage displaying the information about the animal. If the animal doesn't exist or there is no data available, appropriate messages are displayed on the webpage.